Image Optimizer/pyproject.toml
[project]
name = "imageoptimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pillow>=11.2.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "ruff>=0.11.5",
    "tqdm>=4.67.1",
    "webp>=0.4.0",
]
